refactor: move appropriate classes from Admin\Private to UserGroup
