+++ library/Admin/Private/Config/UserGroupRestrictionConfig.php
@@ -4,11 +4,11 @@ namespace Municipio\Admin\Private\Config;
 
 class UserGroupRestrictionConfig implements UserGroupRestrictionConfigInterface
 {
-  /**
-   * Get the user group visibility meta key.
-   *
-   * @return string
-   */
+    /**
+     * Get the user group visibility meta key.
+     *
+     * @return string
+     */
     public function getUserGroupVisibilityMetaKey(): string
     {
         return 'user-group-visibility';
